// src/data/types.ts
export type ServantClass =
  | 'saber'
  | 'archer'
  | 'lancer'
  | 'rider'
  | 'caster'
  | 'assassin'
  | 'berserker'
  | 'ruler'
  | 'avenger'
  | 'alterEgo'
  | 'moonCancer'
  | 'pretender'
  | 'foreigner'
  | 'shielder'
  | 'beast';

export type ServantAttribute = 'sky' | 'earth' | 'human' | 'star' | 'beast';

export type CardColor = 'buster' | 'arts' | 'quick';

export type CardInitial = 'B' | 'A' | 'Q';